workers/oracle_dwarf/requirements.txt
pyelftools==0.31
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings>=2.0
pytest>=7.0
